requirements.txt
gradio
google-generativeai
textblob
deepface
opencv-python
sentence-transformers
faiss-cpu
gradio
